+++ src/main/resources/db/migration/V1_2_4__Specialist.sql
@@ -1,6 +1,6 @@
 CREATE TABLE specialist (
-    id              BIGINT          PRIMARY KEY REFERENCES public.user(id) ON DELETE CASCADE,
-    enterprise_id   BIGINT          NOT NULL,
+    id              INTEGER          PRIMARY KEY REFERENCES public.user(id) ON DELETE CASCADE,
+    enterprise_id   INTEGER         NOT NULL,
 
     CONSTRAINT fk_enterprise
         FOREIGN KEY (enterprise_id) REFERENCES  public.enterprise(id) ON DELETE CASCADE
