+++ src/main/resources/db/migration/V1_3_2__Personal_Account.sql
@@ -1,6 +1,6 @@
 CREATE TABLE public.personal_account (
-    id              BIGINT          PRIMARY KEY REFERENCES account(id) ON DELETE CASCADE,
-    holder_id       BIGINT          NOT NULL,
+    id              INTEGER         PRIMARY KEY REFERENCES account(id) ON DELETE CASCADE,
+    holder_id       INTEGER         NOT NULL,
 
     CONSTRAINT fk_holder
         FOREIGN KEY (holder_id) REFERENCES public.client(id)
